из класса UsersPresenter вынес вложенный класс UsersListPresenter в отдельный класс
