--- app/src/main/java/com/hfad/popularlibrariesrxjava2dagger2moxy/PopularLibrariesExtensions.kt
@@ -1,5 +1,0 @@
-package com.hfad.popularlibrariesrxjava2dagger2moxy
-
-import android.view.View
-
-fun View.click(click: () -> Unit) = setOnClickListener { click() }
