+++ app/src/main/java/com/hfad/popularlibrariesrxjava2dagger2moxy/mvp/presenters/users/UsersPresenter.kt
@@ -2,10 +2,7 @@ package com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.presenters.users
 
 import com.github.terrakok.cicerone.Router
 import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.data.user.GitHubUserRepositoryImpl
-import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.data.user.GitHubUser
 import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.presenters.IScreens
-import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.presenters.IUserListPresenter
-import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.views.UserItemView
 import com.hfad.popularlibrariesrxjava2dagger2moxy.mvp.views.users.UsersView
 import moxy.MvpPresenter
 
@@ -15,21 +12,7 @@ class UsersPresenter(
     private val screens: IScreens
 ) : MvpPresenter<UsersView>() {
 
-    class UsersListPresenter : IUserListPresenter {
-
-        val users = mutableListOf<GitHubUser>()
-
-        override var itemClickListener: ((UserItemView) -> Unit)? = null
-
-        override fun getCount() = users.size
-
-        override fun bindView(view: UserItemView) {
-            val user = users[view.pos]
-            view.setLogin(user.login)
-        }
-    }
-
-     val usersListPresenter = UsersListPresenter()
+    private val usersListPresenter = UsersListPresenter()
 
     override fun onFirstViewAttach() {
         super.onFirstViewAttach()
@@ -48,7 +31,7 @@ class UsersPresenter(
         viewState.updateList()
     }
 
-    fun backPressed() : Boolean {
+    fun backPressed(): Boolean {
         router.exit()
         return true
     }
